Use remoteService's hostname instead of referrer IP
In an environment with an mDNS "reflector" or repeater service (such as a Ubiquiti Unifi network), the `referrer` is not the actual target device but a network device. Therefore, its IP would be e.g. that of a router in the network.

Instead, it should be much safer to simply use the remoteService's hostname and let regular DNS resolution figure out its IP address.